+++ src/Icons/LineIconsTheme.php
@@ -79,6 +79,7 @@ class LineIconsTheme implements IconTheme
         AppShellIcons::DOWNLOAD => 'cloud-download',
         AppShellIcons::FOLDER => 'folder',
         AppShellIcons::FILE => 'empty-file',
+        AppShellIcons::COMMENT => 'comments-alt-2',
     ];
 
     public static function getName(): string