+++ src/Icons/ZmdiIconTheme.php
@@ -79,6 +79,7 @@ class ZmdiIconTheme implements IconTheme
         AppShellIcons::DOWNLOAD => 'cloud-download',
         AppShellIcons::FOLDER => 'folder',
         AppShellIcons::FILE => 'file',
+        AppShellIcons::COMMENT => 'comment',
     ];
 
     public static function getName(): string